master: Updated to avoid addon colition and bin adding . files.

+++ lib/build_bins.py
@@ -1,13 +1,12 @@
 #!/bin/env python
 """Build bins."""
-import calendar
 import json
 import re
 import shutil
 import time
 import os
 from helpers import (paths, mk_clither_custom_dirs, append_to_file, 
-  process_area, get_dir_list, create_symlink, create_directory)
+  process_area, get_dir_list, create_symlink, create_directory, get_epoc_time)
 
 def get_new_path(path, prefix):
     #TODO(xnz): not system portable...
@@ -21,7 +20,7 @@ def _build_paths(config, addon_path):
   # if the paths file does not exists
   # verify that the shellrc has a connection to paths file
   for bin_path in config:
-    bin_path = os.path.join(addon_path, 'bins', bin_path)
+    bin_path = os.path.join(addon_path, bin_path)
     new_path = get_new_path(bin_path, 'addon')
 
     src_dir_path = os.path.join(addon_path, bin_path)
@@ -30,6 +29,9 @@ def _build_paths(config, addon_path):
     file_names = get_dir_list(src_dir_path)
 
     for file_name in file_names:
+      if file_name.startswith('.'):
+        continue
+
       src_full_path = os.path.join(src_dir_path, file_name)
       if os.path.isdir(src_full_path):
         continue
@@ -49,7 +51,7 @@ def _build_paths(config, addon_path):
           continue
 
         create_directory(paths.bin_conflicts_path)
-        epoch_time = str(calendar.timegm(time.gmtime()))
+        epoch_time = get_epoc_time()
         src_base = os.path.basename(src)
         conflict_dst = '{0}_{1}'.format(epoch_time, src_base)
         conflict_dst = os.path.join(paths.bin_conflicts_path, conflict_dst)
@@ -60,27 +62,27 @@ def _override(dir):
 
 def mk_paths_links():
 
-  print('Get content for $PATH, exclued clither_custom')  # not sure how to do this yet
-  sys_bin_paths = '/Users/phuntzinger/.nvm/versions/node/v10.15.0/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/sbin:/usr/local/munki:/Users/phuntzinger/bin'
+  sys_bin_paths = os.environ["PATH"]
   sys_bin_paths = sys_bin_paths.split(':')
-  print('maybe clear out everyting in clither_custom/bin?')
 
   # use this later
-  # log_index = len(str(sys_bin_paths))
+  log_index = len(str(len(sys_bin_paths)))
 
   for index, sys_bin_path in enumerate(sys_bin_paths):
     if sys_bin_path == paths.bin_path:
       # not sure this is right? but ok for now
       continue
 
-    new_path = get_new_path(sys_bin_path, 'path_' + str(index))
+    index = str(index)
+    new_path = get_new_path(sys_bin_path, 'path_' + index.zfill(log_index))
     create_symlink(sys_bin_path, new_path)
 
 def main():
   print('-' * 40)
   print('Start build_bins...')
 
-  shutil.rmtree(paths.bin_path)
+  if os.path.exists(paths.bin_path):
+    shutil.rmtree(paths.bin_path)
   mk_clither_custom_dirs()
   mk_paths_links()
   print('TODO: make sure clither_custom/bin is in $PATH')
